refactor(model): move layer switching from NavigationService onto Subplot
`Context.stepTrace` reached `NavigationService.stepTraceInSubplot`, the last
reason `src/model/` imported `@service/navigation` and so the last violation
of the model rule for this class.

The method never was service work. Its body only touches a `Subplot` and the
`Trace`s inside it: read the current X, step the layer, preserve the position,
notify. Every collaborator is model, and the caller is a model class. It moves
to `Subplot.switchLayer`, which is where the Composite pattern already puts a
subplot-scoped move, and its private `notifyLayerSwitch` moves with it.

Named `switchLayer` rather than `stepTrace` to keep it apart from the
`stepLayer` it calls: `stepLayer` is the silent cursor step, `switchLayer` the
whole switch including the announcement. `stepLayer`'s doc comment already
pointed at the old name, so it now points at the new one.

That leaves `src/service/navigation.ts` with nothing in it, so it goes. The
controller never constructed it and no service referenced it; it was only ever
reachable from the model. `test/service/navigationLayerSwitch.test.ts` moves to
`test/model/layerSwitch.test.ts` and drives the subplot directly — the
assertions are unchanged, so the boundary-announcement behaviour they pin is
verified across the move.
